Refactor Random into a new Bitcoin\Crypto\Random namespace. Working towards deprecating KProvider, just passing a RBGInterface instead
