+++ src/Crypto/Random/Random.php
@@ -1,15 +1,16 @@
 <?php
 
-namespace Bitcoin\Crypto;
+namespace Bitcoin\Crypto\Random;
 
 use Bitcoin\Buffer;
+use Bitcoin\Crypto\Random\RBGInterface;
 use \Bitcoin\Exceptions\RandomBytesFailure;
 
 /**
  * Class Random
  * @package Bitcoin\Random
  */
-class Random
+class Random implements RBGInterface
 {
 
     /**
@@ -18,7 +19,7 @@ class Random
      * @return Buffer
      * @throws RandomBytesFailure
      */
-    public static function bytes($length = 32)
+    public function bytes($length = 32)
     {
         $random = mcrypt_create_iv(32, \MCRYPT_DEV_URANDOM);
